+++ scripts/download_bfd.sh
@@ -37,7 +37,28 @@ SOURCE_URL="https://storage.googleapis.com/alphafold-databases/casp14_versions/b
 BASENAME=$(basename "${SOURCE_URL}")
 
 mkdir --parents "${ROOT_DIR}"
-aria2c "${SOURCE_URL}" --dir="${ROOT_DIR}"
-tar --extract --verbose --file="${ROOT_DIR}/${BASENAME}" \
-  --directory="${ROOT_DIR}"
-rm "${ROOT_DIR}/${BASENAME}"
+
+date_modified=$(curl -I $SOURCE_URL | grep -i last-modified | cut -d " " -f3-5)
+date_file=$(date -r "$ROOT_DIR/$BASENAME" "+%d %b %Y" | cut -d " " -f1-4)
+echo "DATE_MODIFIED: $date_modified"
+echo "DATE_FILE: $date_file"
+
+if [ "$date_modified" == "$date_file" ]; then
+  echo "O database local já está atualizado"
+  exit 1
+else
+  set +e #continua mesmo com erro no aria2c
+  aria2c --allow-overwrite "${SOURCE_URL}" --dir="${ROOT_DIR}" 2> /dev/null
+  
+  if [[ $? -ne 0 ]]; then
+  set -e
+    echo "Erro do download com o aria2, tentando com wget"
+    wget -c -P "${ROOT_DIR}" "https://storage.googleapis.com/alphafold-databases/casp14_versions/${BASENAME}"
+  fi
+
+  tar --extract --verbose --file="${ROOT_DIR}/${BASENAME}" \
+    --directory="${ROOT_DIR}"
+  rm "${ROOT_DIR}/${BASENAME}"
+  touch -d "$date_modified" "${ROOT_DIR}/bfd_metaclust_clu_complete_id30_c90_final_seq.sorted_opt_a3m.ffindex"
+
+fi
