+++ scripts/download_uniprot.sh
@@ -42,14 +42,37 @@ SPROT_BASENAME=$(basename "${SPROT_SOURCE_URL}")
 SPROT_UNZIPPED_BASENAME="${SPROT_BASENAME%.gz}"
 
 mkdir --parents "${ROOT_DIR}"
-aria2c "${TREMBL_SOURCE_URL}" --dir="${ROOT_DIR}"
-aria2c "${SPROT_SOURCE_URL}" --dir="${ROOT_DIR}"
-pushd "${ROOT_DIR}"
-gunzip "${ROOT_DIR}/${TREMBL_BASENAME}"
-gunzip "${ROOT_DIR}/${SPROT_BASENAME}"
 
-# Concatenate TrEMBL and SwissProt, rename to uniprot and clean up.
-cat "${ROOT_DIR}/${SPROT_UNZIPPED_BASENAME}" >> "${ROOT_DIR}/${TREMBL_UNZIPPED_BASENAME}"
-mv "${ROOT_DIR}/${TREMBL_UNZIPPED_BASENAME}" "${ROOT_DIR}/uniprot.fasta"
-rm "${ROOT_DIR}/${SPROT_UNZIPPED_BASENAME}"
-popd
+date_modified=$(curl -I $TREMBL_SOURCE_URL | grep -i last-modified | cut -d " " -f3-5)
+date_file=$(date -r "${ROOT_DIR}/uniprot.fasta" "+%d %b %Y" | cut -d " " -f1-4)
+echo "DATE_MODIFIED: $date_modified"
+echo "DATE_FILE: $date_file"
+
+if [ "$date_modified" == "$date_file" ]; then
+  echo "O database local já está atualizado"
+  exit 1
+else
+    set +e #continua mesmo com erro no aria2c
+
+    aria2c --allow-overwrite "${TREMBL_SOURCE_URL}-" --dir="${ROOT_DIR}"
+    aria2c --allow-overwrite "${SPROT_SOURCE_URL}-" --dir="${ROOT_DIR}"
+    
+
+    if [[ $? -ne 0 ]]; then
+    set -e
+        echo "Erro do download com o aria2, tentando com wget"
+        wget -c -P "${ROOT_DIR}" "${TREMBL_SOURCE_URL}"
+        wget -c -P "${ROOT_DIR}" "${SPROT_SOURCE_URL}"
+    fi
+
+    pushd "${ROOT_DIR}"
+    gunzip "${ROOT_DIR}/${TREMBL_BASENAME}"
+    gunzip "${ROOT_DIR}/${SPROT_BASENAME}"
+
+    # Concatenate TrEMBL and SwissProt, rename to uniprot and clean up.
+    cat "${ROOT_DIR}/${SPROT_UNZIPPED_BASENAME}" >> "${ROOT_DIR}/${TREMBL_UNZIPPED_BASENAME}"
+    mv "${ROOT_DIR}/${TREMBL_UNZIPPED_BASENAME}" "${ROOT_DIR}/uniprot.fasta"
+    rm "${ROOT_DIR}/${SPROT_UNZIPPED_BASENAME}"
+    touch -d "$date_modified" "${ROOT_DIR}/uniprot.fasta"
+    popd
+fi
